test: add tool registration smoke test
